+++ games/maps/js/map/map-selection.js
@@ -1,4 +1,4 @@
-import { OBJECT_LAYER_IDS } from "./map-layers.js?v=2026-08-26.30";
+import { OBJECT_LAYER_IDS } from "./map-layers.js?v=2026-08-26.31";
 
 export function setupSelection(map, { onSelect, isSelectable }) {
   map.on("click", OBJECT_LAYER_IDS, (event) => {